+++ sdk/issuingAuthorization/index.js
@@ -1,0 +1,5 @@
+const issuingAuthorization = require('./issuingAuthorization.js');
+
+exports.parse = issuingAuthorization.parse;
+exports.response = issuingAuthorization.response;
+exports.IssuingAuthorization = issuingAuthorization.IssuingAuthorization;