+++ sdk/issuingBalance/index.js
@@ -1,0 +1,4 @@
+const issuingBalance = require('./issuingBalance.js');
+
+exports.get = issuingBalance.get;
+exports.IssuingBalance = issuingBalance.IssuingBalance;