+++ sdk/issuingBin/index.js
@@ -1,0 +1,5 @@
+const issuingBin = require('./issuingBin.js');
+
+exports.query = issuingBin.query;
+exports.page = issuingBin.page;
+exports.IssuingBin = issuingBin.IssuingBin;