+++ sdk/issuingCard/index.js
@@ -1,0 +1,10 @@
+const issuingCard = require('./issuingCard.js');
+
+exports.log = require('./log');
+exports.create = issuingCard.create;
+exports.get = issuingCard.get;
+exports.query = issuingCard.query;
+exports.page = issuingCard.page;
+exports.update = issuingCard.update;
+exports.delete = issuingCard.delete;
+exports.IssuingCard = issuingCard.IssuingCard;