+++ sdk/issuingInvoice/index.js
@@ -1,0 +1,8 @@
+const issuingInvoice = require('./issuingInvoice.js');
+
+exports.log = require('./log');
+exports.create = issuingInvoice.create;
+exports.get = issuingInvoice.get;
+exports.query = issuingInvoice.query;
+exports.page = issuingInvoice.page;
+exports.IssuingInvoice = issuingInvoice.IssuingInvoice;