+++ sdk/issuingInvoice/log/index.js
@@ -1,0 +1,4 @@
+const log = require('./log.js');
+
+exports.query = log.query;
+exports.page = log.page;