+++ sdk/issuingWithdrawal/index.js
@@ -1,0 +1,7 @@
+const issuingWithdrawal = require('./issuingWithdrawal.js');
+
+exports.create = issuingWithdrawal.create;
+exports.get = issuingWithdrawal.get;
+exports.query = issuingWithdrawal.query;
+exports.page = issuingWithdrawal.page;
+exports.IssuingWithdrawal = issuingWithdrawal.IssuingWithdrawal;