+++ CHANGELOG.md
@@ -14,6 +14,15 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 ## [Unreleased]
 - CreditNote resource for money lending with Stark Infra's endorsement
+- IssuingAuthorization resource for Sub Issuers
+- IssuingBalance resource for Sub Issuers
+- IssuingBin resource for Sub Issuers
+- IssuingCard resource for Sub Issuers
+- IssuingHolder resource for Sub Issuers
+- IssuingInvoice resource for Sub Issuers
+- IssuingPurchase resource for Sub Issuers
+- IssuingTransaction resource for Sub Issuers
+- IssuingWithdrawal resource for Sub Issuers
 
 ## [0.0.1] - 2022-03-18
 ### Added